[SSHD-1020] Added a default NIO2-READ-TIMEOUT value for tests
